Graph and LinearRegressionModel Objects split up
The Graph object has had all functions related to linear regression
removed. Now, its sole function is to be a graph.

The linear regression functions were moved to a LinearRegressionModel
object. This object can be initialized and added to the Graph object as
a "dependency.' The graph object will act as a subject with dependencies
as observers.