Cost Function Graph plots beta and RSS
A cost function graph has been added that plots the current beta and
RSS, yielding a parabolic function. This was accomplished by making a
graph that observes the linearRegressionModel. When the
linearRegressionModel sees a mousemove, it calls addPoint() on its
observers. This then adds a point to the costFunctionGraph.

Currently, there is a bug because the original points are not plotting
in the linearRegressionGraph.